api/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
hvac==0.9.5
google-auth==1.6.3
google-auth-oauthlib==0.4.0
google-auth-httplib2==0.0.3
